arm64: tegra: fix build warning for hotplug driver
-tegra_suspend_in_progress(): function declaration
 isn't a prototype
-cast from pointer to integer of different size

Bug 1211919

Reviewed-on: http://git-master/r/411539

+++ arch/arm64/mach-tegra/hotplug-denver.c
@@ -79,7 +79,8 @@ noinline int mca_cpu_callback(struct notifier_block *nfb,
 				unsigned long action, void *hcpu)
 {
 	if(action == CPU_ONLINE || action == CPU_ONLINE_FROZEN)
-		smp_call_function_single((int) hcpu, setup_mca, NULL, 1);
+		smp_call_function_single((int)(uintptr_t) hcpu,
+						setup_mca, NULL, 1);
 	return 0;
 }
